[shape_poly] Add support for shape polymorphism for LU on CPU.
We support polymorphism only on the batch sizes for now.
This did not require any changes to C++ custom call targets, but
needed to move some logic from the jaxlib to JAX, regarding
the computation of shapes and types.

Also added backwards compatibility tests for the LU custom calls
for CPU, and improved the checking of LU results by checking
the invariant for the result as opposed to checking goldens.

PiperOrigin-RevId: 540464911